added train.py, dataset.py
only stagewise and pretrain refactored till now
