+++ semantic_segmentation/code/core.py
@@ -1,0 +1,18 @@
+from comet_ml import Experiment
+import torch
+import torch.nn as nn
+from torchvision import transforms
+from torch.utils.data import DataLoader
+import torch.nn.functional as F
+import models
+import argparse
+from helper import *
+from args import get_args
+from features import get_features
+from pathlib import Path
+#
+#
+#
+# class Camv
+# camvid_tfms =
+
