+++ semantic_segmentation/code/eval_trad_kd.py
@@ -1,51 +1,16 @@
-%matplotlib inline
 import torch
-import torch.nn as nn
-from torchvision import transforms
 from torch.utils.data import DataLoader
 import torch.nn.functional as F
 import models
-import matplotlib.pyplot as plt
+from dataset import CamVid
 from helper import *
 torch.cuda.set_device(0)
 
-classes = ['sky', 'building', 'pole', 'road', 'pavement', 'tree', 'signsymbol', 'fence', 'car', 'pedestrian', 'bicyclist', 'unlabelled']
+num_classes = 12
 DATA_DIR = '../data/CamVid/'
-
-x_train_dir = os.path.join(DATA_DIR, 'train')
-y_train_dir = os.path.join(DATA_DIR, 'trainannot')
-
-x_valid_dir = os.path.join(DATA_DIR, 'val')
-y_valid_dir = os.path.join(DATA_DIR, 'valannot')
-
-x_test_dir = os.path.join(DATA_DIR, 'test')
-y_test_dir = os.path.join(DATA_DIR, 'testannot')
-
-transform = transforms.Compose([
-    transforms.ToTensor(),
-    transforms.Normalize(mean = [0.41189489566336, 0.4251328133025, 0.4326707089857], std = [0.27413549931506, 0.28506257482912, 0.28284674400252])
-])
-
-train_dataset = CamVid(
-    x_train_dir,
-    y_train_dir,
-    classes = classes,
-    transform = transform
-)
-
-valid_dataset = CamVid(
-    x_valid_dir,
-    y_valid_dir,
-    classes = classes, 
-    transform = transform
-)
-
-test_dataset = CamVid(
-    x_test_dir,
-    y_test_dir,
-    classes = classes, 
-    transform = transform
-)
+train_dataset = CamVid(DATA_DIR, mode='train')
+valid_dataset = CamVid(DATA_DIR, mode='val')
+test_dataset = CamVid(DATA_DIR, mode='test')
 
 trainloader = DataLoader(train_dataset, batch_size = 8, shuffle = True, drop_last = True)
 valloader = DataLoader(valid_dataset, batch_size = 1, shuffle = False)
